requirements.txt
dnspython==2.2.0
requests